fixing issues
git-svn-id: https://svn-bccfr.uis.georgetown.edu/svn/gdoc/webapp/trunk@1044 7e94895d-eddb-4489-9ea7-1067dfd99f16

+++ grails-app/services/CleanupService.groovy
@@ -21,8 +21,10 @@ class CleanupService implements serviceinterfaces.SessionCleanerServiceInt {
 			def tbSize = analyses.size()
 			analyses.each{
 				def analysis = SavedAnalysis.get(it)
-				if(savedAnalysisService.deleteAnalysis(analysis.id))
-				log.debug "deleted $tbSize temp analyses for $author.loginName";
+				if(analysis) {
+					if(savedAnalysisService.deleteAnalysis(analysis.id))
+						log.debug "deleted $tbSize temp analyses for $author.loginName";
+				}
 			}
 			
 		}
